requirements.txt
bcrypt==4.0.1
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.2
simple-websocket==0.9.0
six==1.16.0
socketIO-client-nexus==0.7.6
speedtest-cli==2.1.3
SQLAlchemy==2.0.4
typing_extensions==4.5.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.14.0
zope.event==4.6
zope.interface==5.5.2
